Binance api/requirements.txt
Backtesting
binance-futures-connector
openpyxl
pandas
python-binance
ta